alterações usuarios

+++ controller/filmeExcluir.php
@@ -1,11 +1,14 @@
 ﻿<?php
+  require_once('../model/autenticar.php');
   require_once('../model/conexao.php');
   $filmeDelete = file_get_contents('php://input');
   $filmeMatriz = json_decode($filmeDelete, true);
   $id = (isset($filmeMatriz["id"]) && $filmeMatriz["id"] != null) ? $filmeMatriz["id"] : null;
-  $resposta["erro"] = false;$resposta["msgErro"] = "";
-  $resposta["msgSucesso"] = "";$resposta["dados"] = null;
-  if( $id != null ){
+  $resposta["erro"] = false;
+  $resposta["msgErro"] = "";
+  $resposta["msgSucesso"] = "";
+  $resposta["dados"] = null;
+  if( $id != null && $autenticado===true){
     try {
         $sql = "DELETE FROM filmes_assistidos WHERE id=?";
         $stmt = $conexao->prepare($sql);
@@ -13,8 +16,8 @@
         $stmt->execute();
         $resposta["msgSucesso"] = "Filme de id $id excluído com sucesso!";
     }catch(PDOException $e) {
-      $resposta["erro"] = true;
-      $resposta["msgErro"] = "Erro: ".$e->getMessage();
+        $resposta["erro"] = true;
+        $resposta["msgErro"] = "Erro: ".$e->getMessage();
     }finally{
         echo json_encode($resposta);  
     }
